[mpact][compiler] move sample models into centralized models directory
Good idea by Yinying, so that tests and benchmark can draw from our
sample model pool, without having to repeat the PyTorch code all
over the place
